GH-1171: extract compare/ into internal sub-package
Move compare.go and testloader.go (plus tests) into
pkg/orchestrator/internal/compare/. The root package retains thin
delegation wrappers and type aliases so the public API is unchanged.

External dependencies (logf, binGit, binGo, gitWorktreeRemove) are
injected via a Deps struct, breaking the direct coupling to the
orchestrator package. A Comparer interface defines the sub-package
boundary per the updated go-style constitution.
